+++ packages/config/src/chains/index.ts
@@ -1,8 +1,8 @@
 import { notUndefined } from '@l2beat/shared-pure'
 
-import type { ChainConfig } from '../common/ChainConfig'
 import { layer2s } from '../projects/layer2s'
 import { layer3s } from '../projects/layer3s'
+import type { ChainConfig } from '../types'
 import { avalanche } from './avalanche'
 import { bsc } from './bsc'
 import { celo } from './celo'
