+++ packages/config/src/chains/polygonpos.ts
@@ -1,6 +1,6 @@
 import { EthereumAddress } from '@l2beat/shared-pure'
 
-import type { ChainConfig } from '../common/ChainConfig'
+import type { ChainConfig } from '../types'
 
 export const polygonpos: ChainConfig = {
   name: 'polygonpos',
