+++ packages/config/src/projects/layer2s/templates/polygonCDKStack.ts
@@ -10,7 +10,6 @@ import {
 import { ethereum } from '../../../chains/ethereum'
 import {
   CONTRACTS,
-  type ChainConfig,
   DA_BRIDGES,
   DA_LAYERS,
   DA_MODES,
@@ -40,6 +39,7 @@ import {
 } from '../../../common'
 import { formatDelay, formatExecutionDelay } from '../../../common/formatDelays'
 import { ProjectDiscovery } from '../../../discovery/ProjectDiscovery'
+import type { ChainConfig } from '../../../types'
 import { Badge, type BadgeId, badges } from '../../badges'
 import type { DacDaLayer } from '../../da-beat/types'
 import { getStage } from '../common/stages/getStage'
