+++ packages/config/src/projects/layer2s/templates/underReview.ts
@@ -1,13 +1,13 @@
 import { ProjectId, type UnixTime } from '@l2beat/shared-pure'
 import {
   CONTRACTS,
-  type ChainConfig,
   type ScalingProjectDisplay,
   type ScalingProjectEscrow,
   type ScalingProjectTransactionApi,
   TECHNOLOGY,
   UNDER_REVIEW_RISK_VIEW,
 } from '../../../common'
+import type { ChainConfig } from '../../../types'
 import type { BadgeId } from '../../badges'
 import type { Layer3 } from '../../layer3s'
 import type { Layer2, Layer2Display } from '../types'
